requirements.txt
asgiref==3.2.10
djangorestframework==3.12.4
Django==3.1.6
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
